Add -count option to limit the number of messages
Add a -count parameter to optionally limit the output to a specific
number of messages per worker.

This could be handy for testing when one wants to confirm that all
messages are being delivered. Can send a specific large number of
messages and count how many arrive at the endpoint.